requirements.txt
Django>=5.0.1,<5.0.2
# djangorestframework>=3.12.4,<3.13.0
django-environ
psycopg2>=2.9.2,<2.10.0 #only run it in EC2 - Mac M1 don't need it
psycopg2-binary
Pillow # for files upload management
types-Pillow
yapf
requests
types-requests
pylint
mypy
feedparser
django-stubs>=1.9.0<1.10.0
django-storages
google-cloud-storage
google-cloud-logging
google-cloud-secret-manager
beautifulsoup4
selenium>=4.8.2
algoliasearch>=2.0,<3.0
djangorestframework
webdriver-manager>=4.0.1
django-markdownify
belorthography>=1.0.0
django-debug-toolbar
polib
